requirements.txt
urllib3>=1.26.5
pandas
pandas-ta
numpy
requests==2.25.0
statsmodels
matplotlib==3.3.3
mock==4.0.3
pytest==6.2.2
pytest-mock==3.6.1
responses==0.13.3
pyyaml==5.4.1
websocket-client==0.59.0
websockets==9.1
flask
python-telegram-bot==13.7
tradingview-ta
bottleneck
dash
dash-daq
dash-bootstrap-components
rich
